Car Launcher: fix for Car Launcher UI
This patch is for fixing Car Launcher UI.
After completing boot, the navigation status bar
at the bottom is not all visible and when the
user clicks the Home button, the menu is all visible.

The cause of this problem is that the UI configuration
is not normal because the Car Launcher app is running
in Direct Boot Mode, not provisioning in DefaultActivity.

With non-direct boot mode, it is normal because
DefaultActivity is launched after Boot completed
and the Car Launcher is executed after provisioning.
Also, It has no effect on boot time.

Signed-off-by: chungho.shin <chungho.shin@nexell.co.kr>
Change-Id: I79609cdaf636dc712b3a1e9cffb25aef98f69123
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/android_platform_packages_apps_Car_Launcher/+/489084
Reviewed-by: allan park <allan.park@nexell.co.kr>
Tested-by: allan park <allan.park@nexell.co.kr>